Id type changed to long.

+++ src/main/java/server/filestorm/model/type/CustomSession.java
@@ -5,12 +5,12 @@ import java.util.Map;
 public class CustomSession {
 
     private Map<String, Object> claims;
-    private Integer user_id;
+    private Long user_id;
     private String username;
 
     public CustomSession(Map<String, Object> claims) {
         this.claims = claims;
-        this.user_id = (Integer) this.claims.get("id");
+        this.user_id = (Long) this.claims.get("id");
         this.username = (String) this.claims.get("username");
     }
 
@@ -18,7 +18,7 @@ public class CustomSession {
         return claims;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return user_id;
     }
